Update move backup endpoint to product class (#41730)
* Move backup endpoint to Backup product class

* changelog

* Add visibility to constants in Initializer

* Add test for registering endpoint

* Remove new test
