+++ projects/packages/my-jetpack/changelog/update-move-backup-endpoint-to-product-class
@@ -1,0 +1,4 @@
+Significance: patch
+Type: changed
+
+Move backup endpoint to product class
